requirements.txt
flask
Flask-SQLAlchemy
Flask-WTF
flask-login
flask-admin
Flask-Mail
python-dotenv
gunicorn
psycopg2